早回し関連OKOKOK
git-svn-id: file:///home/gecchi/repos@2900 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafCore/include/jp/ggaf/core/util/GgafUtil.h
@@ -81,7 +81,7 @@ public:
      * @return ハッシュ値
      */
     static UINT32 easy_hash(const char* str) {
-        unsigned long hash = 5381;
+        UINT32 hash = 5381;
         char c;
         while (c = *str++) { //strの\0までループ （演算子 "==" と間違えていません）
             hash = ((hash << 5) + hash) + c; // hash * 33 + c  33倍してます
